requirements.txt
python-telegram-bot==22.0
openai==0.27.0
requests==2.28.1
redis==4.5.1